拆分详情页面各组件
