+++ core/infosphere-console/src/views/layouts/book/DynamicContainer.vue
@@ -7,22 +7,18 @@ import { computed, defineComponent } from 'vue'
 import { useRoute } from 'vue-router'
 import BookContainer from '@/views/layouts/book/BookContainer.vue'
 import SettingContainer from '@/views/layouts/book/SettingContainer.vue'
-import LayoutContainer from '@/views/layouts/basic/LayoutContainer.vue'
 import WriterContainer from '@/views/layouts/book/WriterContainer.vue'
+import InfoContainer from '@/views/layouts/book/InfoContainer.vue'
 
 export default defineComponent({
   name: 'DynamicContainer',
-  components: {
-    BookContainer,
-    SettingContainer
-  },
   setup()
   {
     const route = useRoute()
 
     const currentComponent = computed(() => {
-      if (route.path.startsWith('/book/info')) {
-        return LayoutContainer
+      if (route.path.startsWith('/book/info') || route.path.startsWith('/book/access')) {
+        return InfoContainer
       }
       if (route.path.startsWith('/book/setting')) {
         return SettingContainer
